MOVE-4866 MOVE-4880 oppgradere til msh-client 1.4 plus endepunkt for å hente meldinger per herId
