+++ nhn-adapter-app/pom.xml
@@ -39,7 +39,7 @@
         <java.version>21</java.version>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <helseid.version>1.0.1</helseid.version>
-        <msh-client.version>1.1.0</msh-client.version>
+        <msh-client.version>1.4.0</msh-client.version>
         <kotlin.version>2.2.0</kotlin.version>
         <kotest.version>5.9.1</kotest.version>
         <kotest-spring.version>1.3.0</kotest-spring.version>
