+++ nhn-adapter-app/src/main/kotlin/no/difi/meldingsutveksling/nhn/adapter/handlers/InHandler.kt
@@ -13,9 +13,11 @@ import no.difi.meldingsutveksling.nhn.adapter.crypto.NhnTrustStore
 import no.difi.meldingsutveksling.nhn.adapter.crypto.Signer
 import no.difi.meldingsutveksling.nhn.adapter.model.EncryptedFagmelding
 import no.difi.meldingsutveksling.nhn.adapter.model.SerializableApplicationReceiptInfo
+import no.difi.meldingsutveksling.nhn.adapter.model.toInMessage
 import no.difi.meldingsutveksling.nhn.adapter.model.toSerializable
 import no.ks.fiks.nhn.msh.Client
 import no.ks.fiks.nhn.msh.HelseIdTokenParameters
+import no.ks.fiks.nhn.msh.IncomingBusinessDocument
 import no.ks.fiks.nhn.msh.MultiTenantHelseIdTokenParameters
 import no.ks.fiks.nhn.msh.RequestParameters
 import org.springframework.http.MediaType
@@ -76,4 +78,46 @@ object InHandler {
 
         return ServerResponse.ok().contentType(MediaType.APPLICATION_JSON).bodyValueAndAwait(signedReceipts)
     }
+
+    suspend fun incomingMessages(request: ServerRequest, mshClient: Client): ServerResponse {
+        val herId2 =
+            try {
+                request.pathVariable("herId2").toInt()
+            } catch (e: NumberFormatException) {
+                throw IllegalArgumentException("Message id is wrong format", e)
+            }
+        val onBehalfOf = request.queryParamOrNull("onBehalfOf")
+        val requestParameters =
+            onBehalfOf?.let { onBehalfOf ->
+                RequestParameters(HelseIdTokenParameters(MultiTenantHelseIdTokenParameters(onBehalfOf)))
+            } ?: throw IllegalArgumentException("On behalf of organisation is not provided.")
+
+        // vi bruker status endepunkt for å hente Apprec
+        val notApprecMessages =
+            mshClient.getMessagesWithMetadata(herId2, requestParameters).filter { it.isAppRec == false }
+        val inMessages = notApprecMessages.map { it.toInMessage() }
+
+        return ServerResponse.ok().contentType(MediaType.APPLICATION_JSON).bodyValueAndAwait(inMessages)
+    }
+
+    suspend fun readMessages(request: ServerRequest, mshClient: Client): ServerResponse {
+        val messageId =
+            try {
+                UUID.fromString(request.pathVariable("messageId"))
+            } catch (e: NumberFormatException) {
+                throw IllegalArgumentException("Message id is wrong format", e)
+            }
+
+        val onBehalfOf = request.queryParamOrNull("onBehalfOf")
+        val requestParameters =
+            onBehalfOf?.let { onBehalfOf ->
+                RequestParameters(HelseIdTokenParameters(MultiTenantHelseIdTokenParameters(onBehalfOf)))
+            } ?: throw IllegalArgumentException("On behalf of organisation is not provided.")
+
+        val businessDokument: IncomingBusinessDocument = mshClient.getBusinessDocument(messageId, requestParameters)
+
+        return ServerResponse.ok().bodyValueAndAwait(businessDokument)
+    }
+
+    suspend fun markAsRead(request: ServerRequest): ServerResponse = ServerResponse.ok().bodyValueAndAwait("")
 }
